Fix bug

+++ app/views/jurusan/index.php
@@ -23,21 +23,15 @@
         </div>
 
         <ul class="list-group">
-    <?php foreach ($data['jrs'] as $jrs) : ?>
-    <li class="list-group-item">
-        <?= $jrs['nama_jurusan'] ?>
+            <?php foreach ($data['jrs'] as $jrs) : ?>
+         <li class="list-group-item">
+            <?= $jrs['nama_jurusan'] ?>
         <div class="d-flex justify-content-end align-items-end">
-            <a href="<?= BASEURL; ?>/jurusan/hapusJurusan/<?= $jrs['id']; ?>" 
-               class="btn btn-danger text-white me-2"
-               onclick="return confirm('Apakah Anda yakin ingin menghapus jurusan ini?');">
-               Delete
-            </a>
+            <a href="<?= BASEURL;?>/jurusan/hapusJurusan" class="btn btn-danger text-white me-2 deleteMhs2" onclick="confirmDeleteB(<?= $jrs ['id'];?>, event)">Delete</a>
         </div>
-    </li>
-    <?php endforeach ?>
-</ul>
-
-
+        </li>
+            <?php endforeach; ?>
+        </ul>
             </div>
         </div>
     </div>
